+++ web/playground/__tests__/e2e/common/auth.ts
@@ -27,7 +27,11 @@ export async function authLogin(page: Page) {
   const startX = actionBoundingBox.x + actionBoundingBox.width / 2; // div 中心的 x 坐标
   const startY = actionBoundingBox.y + actionBoundingBox.height / 2; // div 中心的 y 坐标
 
-  const targetX = startX + sliderCaptchaBox.width + actionBoundingBox.width; // 向右拖动容器的宽度
+  const targetX =
+    sliderCaptchaBox.x +
+    sliderCaptchaBox.width -
+    actionBoundingBox.width / 2 -
+    2;
   const targetY = startY; // y 坐标保持不变
 
   // 模拟鼠标拖动
@@ -39,8 +43,10 @@ export async function authLogin(page: Page) {
   // 在拖动后进行断言，检查action是否在预期位置,
   const newActionBoundingBox = await sliderCaptchaAction.boundingBox();
   expect(newActionBoundingBox?.x).toBeGreaterThan(actionBoundingBox.x);
+  await expect(sliderCaptcha).toContainText(/Passed|验证通过/);
 
   // 到这里已经校验成功，点击进行登录
   await page.waitForTimeout(300);
   await page.getByRole('button', { name: 'login' }).click();
+  await expect(page).toHaveURL(/\/dashboard\/overview/);
 }